+++ packages/tools/can-i-use/rollup.config.js
@@ -3,25 +3,14 @@ import commonjs from '@rollup/plugin-commonjs';
 import swc from 'rollup-plugin-swc';
 import { minify } from 'rollup-plugin-esbuild';
 import cleaner from 'rollup-plugin-cleaner';
-// import babel from '@rollup/plugin-babel';
-import image from '@rollup/plugin-image';
-// 使rollup可以使用postCss处理样式文件less、css等
-import postcss from 'rollup-plugin-postcss';
-// 处理css定义的变量
-import simplevars from 'postcss-simple-vars';
-// 处理less嵌套样式写法
-import nested from 'postcss-nested';
-// autoPrefix
-import autoprefixer from 'autoprefixer';
 import progress from 'rollup-plugin-progress';
-import gStyle from '@montagejs/rollup-plugin-generate-style';
 
-const extensions = ['.js', '.jsx', '.ts', '.tsx'];
+const extensions = ['.js', '.ts'];
 
-const external = ['react', '@waline/client', 'classnames', 'react/jsx-runtime'];
+const external = [];
 
 export default {
-    input: './src/index.tsx',
+    input: './src/index.ts',
     output: [
         {
             file: './lib/index.js',
@@ -38,7 +27,7 @@ export default {
             file: './dist/index.js',
             sourcemap: false,
             format: 'umd',
-            name: 'ReactWalineClient',
+            name: 'CanIUse',
         },
     ],
     external,
@@ -51,23 +40,6 @@ export default {
         cleaner({
             targets: ['./es/', './lib/', '.dist'],
         }),
-        postcss({
-            plugins: [simplevars(), nested(), autoprefixer()],
-            minimize: true,
-            extract: 'style/index.css',
-            // 处理.css和.less文件
-            extensions: ['.less', '.css'],
-            use: [
-                [
-                    'less',
-                    {
-                        javascriptEnabled: true,
-                    },
-                ],
-            ],
-        }),
-        gStyle(),
-        image(),
         // swc 提供 jsx runtime
         swc({
             exclude: ['.*.js$', '.*.map$', '.*.d.ts$'],
@@ -83,22 +55,9 @@ export default {
                     syntax: 'typescript',
                     tsx: true,
                 },
-                transform: {
-                    react: {
-                        runtime: 'automatic',
-                        useBuiltins: true,
-                        development: false,
-                    },
-                },
                 target: 'es2016',
             },
         }),
-        // babel import 实现按需加载
-        // babel({
-        //     extensions,
-        //     exclude: /node_modules/,
-        //     configFile: './babel.lazy.config.js',
-        // }),
         // esbuild 压缩代码
         minify(),
     ],