+++ packages/tools/can-i-use/src/helper.ts
@@ -1,0 +1,1 @@
+export const kindOf = (obj: any) => Object.prototype.toString.call(obj).slice(1, -8).toLowerCase();