+++ packages/plugins/rollup-plugin-generate-style/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @montagejs/rollup-plugin-generate-style
 
+## 0.0.4
+
+### Patch Changes
+
+- [`e8f797a`](https://github.com/bigbigDreamer/montage/commit/e8f797a) Thanks [@bigbigDreamer](https://github.com/bigbigDreamer)! - fix: add some package fileds
+
 ## 0.0.3
 
 ### Patch Changes